Update CORS policy, change price from int to float

+++ migrations/sqls/20221008131825-products-table-up.sql
@@ -3,5 +3,5 @@ CREATE TABLE IF NOT EXISTS products (
 	name VARCHAR(100) NOT NULL,
 	url TEXT DEFAULT NULL,
 	description TEXT DEFAULT NULL,
-	price INT NOT NULL
+	price FLOAT(2) NOT NULL
 );
